+++ src/main/java/com/phuclq/student/service/AttachmentService.java
@@ -2,6 +2,7 @@ package com.phuclq.student.service;
 
 import com.phuclq.student.dto.AttachmentDTO;
 import com.phuclq.student.dto.RequestFileDTO;
+import io.swagger.models.auth.In;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
@@ -13,7 +14,7 @@ public interface AttachmentService {
   Long createListAttachmentsFromBase64S3(List<RequestFileDTO> files, Integer requestId,Integer loginId)
       throws IOException;
 
-  AttachmentDTO getAttachmentByIdFromS3(Long id,String fileType, HttpServletRequest request) throws IOException;
+  AttachmentDTO getAttachmentByIdFromS3(Integer id,String fileType, HttpServletRequest request) throws IOException;
 
   AttachmentDTO getAttachmentByRequestIdFromS3(Integer requestId,String fileType)
       throws IOException;
